v1.16.0 (#509)
* fix: font colors for total quantity column in overview (#483)

* feat: add duplicate sku/barcode validation to product import schema (#484)

* feat: add duplicate sku/barcode validation to product import schema

* chore: add english translation

* chore(db): add nl_default_placement table in migrations

* feat: default placement functionality + ui

* fix: logic for showing cols and filters in overview

* fix: requested changes

* Feat/web location move (#487)

* feat: wip locations move

* feat: first take on move inventory logic

* fix: pr changes

* feat: added translations and error messages

* fix: toast titles, small ux changes

* feat: show sku in add products popover

* fix: disable button when form is submitting

* fix: spinner on button when location is changed.

form.handleSubmit and formState.isSubmitting did not work :(

* fix: locations move for moderator

* feat: can now set location to primary even if relation is not yet in
database

* feat: hide upload file ui if customer is not basis or up

* fix: allow users to upload and delete files so it aligns with ui
permissions

* fix: bug where users was not able to upload files from firefox

we used `tobase64` before which is still a relatively new feature
so browsers may not have the same implementation e.g. chrome and
firefox. firefox took in a options object and chrome just took in
a string.

* feat: disallow lite customers from uploading files

* feat: add label to default placement (#496)

* feat: add checks if attachment exists and belongs to customer

* Feat/delete restock (#499)

* feat: modal for deleting reorder

* feat: server action to delete reorder

* feat: translations for deleting reorder

* fix: moved delete modal to its own file

* feat: data function to fetch inventories by placement id

* feat: ui + server action

* fix: remove successful inventories when bulk fails

* feat: add to skeleton comp

* fix: include translation in fetch inventories action

* fix: take into account if customer has disabled reference

* fix: take into account if customer has disabled reference

* Feat/redirect query param (#508)

* feat(sign-in-page): add redirect query param

* feat: use current path name to redirect if auth fails

* feat(site) : add withAuth to all pages

* Feat/batch rework (#506)

* feat: add useBatch to db product table

* feat: add batch setting to product details page

* chore: remove useBatch column from customerSettings table

* chore: remove usages of useBatch on customer settings

* chore: return hasPermissionByPlan 'pro' as useBatch in setting endpoints

* chore: add .envrc to .gitignore

* feat: batch selection based on product settings

* feat: batch selection based on product setting in inv modals

* feat: qol changes for inv modals

* fix: settings endpoints return use batch based on num of batch products

* fix: missing useBatch after rebase

* Feat/batch lifetimes (#503)

* fix: not as aggressive filter on place and batch in update inv modal

* feat: indicator for batch/place in overview page + move/upt inv modal

* chore: QOL changes for overview page

* feat: move inventories when disabling batch (#505)

* feat(fe): add dialog for disabling batch registrations

* feat(be): move inventories to default batch on disable

* fix: hide batch setting on product details page for non-pro customers

* fix: requested changes

* refactor(oversigt): extract giant then block to separate functions

* refactor(oversigt): remove 1 loop of filtering of inventories

* fix: remove console.log

* fix: missing translation on history logs for batch disable invetory move

* feat(inv-service): add function to get default batches

* refactor(move-inv-modal): remove 1 product filter loop

---------

+++ src/app/[lng]/(site)/varer/produkter/[id]/actions.ts
@@ -1,14 +1,18 @@
 'use server'
 
 import { serverTranslation } from '@/app/i18n'
+import { fallbackLng, strIsI18NLanguage } from '@/app/i18n/settings'
 import { attachmentRefTypeValidation } from '@/data/attachments'
+import { hasPermissionByPlan } from '@/data/user.types'
 import { adminAction, editableAction } from '@/lib/safe-action'
 import { ActionError } from '@/lib/safe-action/error'
 import { attachmentService } from '@/service/attachments'
 import { fileService } from '@/service/file'
 import { inventoryService } from '@/service/inventory'
+import { locationService } from '@/service/location'
 import { productService } from '@/service/products'
 import { suppliersService } from '@/service/suppliers'
+import { revalidatePath } from 'next/cache'
 import { z } from 'zod'
 
 const createAttachmentValidation = z.object({
@@ -30,30 +34,40 @@ const uploadFileValidation = z.object({
   body: z.string(),
 })
 
-export const uploadFileAction = adminAction
+export const uploadFileAction = editableAction
   .schema(uploadFileValidation)
-  .action(async ({ parsedInput: { key, type, body } }) => {
+  .action(async ({ parsedInput: { key, type, body }, ctx: {customer, lang} }) => {
+	const { t } = await serverTranslation(lang, 'produkter')
+
+	if (customer && !hasPermissionByPlan(customer.plan, 'basis')) {
+		throw new ActionError(t("details-page.server.attachment-not-allowed"))
+	}
+
     const b = Buffer.from(body, 'base64url')
     const arraybuffer = new Uint8Array(b)
 
     return await fileService.upload({ key, mimeType: type, body: arraybuffer })
   })
 
-export const createAttachmentAction = adminAction
+export const createAttachmentAction = editableAction
   .schema(createAttachmentValidation)
-  .action(async ({ parsedInput, ctx }) => {
-    const { t } = await serverTranslation(ctx.lang, 'produkter')
+  .action(async ({ parsedInput, ctx: {user, customer, lang} }) => {
+	const { t } = await serverTranslation(lang, 'produkter')
+
+	if (customer && !hasPermissionByPlan(customer.plan, 'basis')) {
+		throw new ActionError(t("details-page.server.attachment-not-allowed"))
+	}
 
     const newAttachment = await attachmentService.create({
-      customerID: ctx.user.customerID,
+      customerID: user.customerID,
       refDomain: parsedInput.refType,
       refID: String(parsedInput.refID),
       name: parsedInput.name,
       type: parsedInput.type,
       key: parsedInput.key,
       url: parsedInput.url,
-      userID: ctx.user.id,
-      userName: ctx.user.name,
+      userID: user.id,
+      userName: user.name,
     })
 
     if (!newAttachment) {
@@ -66,10 +80,20 @@ export const createAttachmentAction = adminAction
     }
   })
 
-export const deleteAttachmentAction = adminAction
+export const deleteAttachmentAction = editableAction
   .schema(deleteAttachmentValidation)
   .action(async ({ parsedInput, ctx }) => {
     const { t } = await serverTranslation(ctx.lang, 'produkter')
+
+	const attachment = await attachmentService.getByID(parsedInput.id)
+	if (!attachment) {
+      throw new ActionError(t('details-page.server.file-not-found'))
+	}
+
+	if (attachment.customerID != ctx.user.customerID) {
+      throw new ActionError(t('details-page.server.file-not-yours'))
+	}
+
     const didDelete = await attachmentService.deleteByID(parsedInput.id)
     if (!didDelete) {
       throw new ActionError(t('details-page.server.file-not-deleted'))
@@ -142,3 +166,86 @@ export const fetchSuppliersAction = editableAction.action(
     return suppliers
   },
 )
+
+export const upsertDefaultPlacement = adminAction
+	.schema(
+		z.object({
+			productID: z.coerce.number(),
+			placementID: z.coerce.number(),
+			locationID: z.coerce.string(),
+		}),
+	)
+	.action(
+		async ({
+			parsedInput: { productID, placementID, locationID },
+			ctx: { user, lang },
+		}) => {
+			await inventoryService.upsertDefaultPlacement(
+				[productID, placementID, locationID],
+				user.customerID,
+				user.id,
+				'web',
+				strIsI18NLanguage(lang) ? lang : fallbackLng,
+			)
+
+			revalidatePath(`${lang}/varer/produkter/${productID}`)
+		},
+	)
+
+export const deleteDefaultPlacement = adminAction
+	.schema(
+		z.object({
+			productID: z.coerce.number(),
+			placementID: z.coerce.number(),
+			locationID: z.coerce.string(),
+		}),
+	)
+	.action(
+		async ({
+			parsedInput: { productID, placementID, locationID },
+			ctx: { lang }
+		}) => {
+			const { t } = await serverTranslation(lang, 'produkter', {keyPrefix: 'details-page.inventory.remove-default-placement-dialog'})
+			console.log(`Remove default placement: [${productID}, ${placementID}, ${locationID}]`)
+
+			const success = await inventoryService.deleteDefaultPlacement([productID, placementID, locationID])
+
+			if (!success) {
+				throw new ActionError(t('toasts.default-placement-not-removed'))
+			}
+
+			revalidatePath(`${lang}/varer/produkter/${productID}`)
+		},
+	)
+
+export const deleteReorderAction = editableAction
+  .metadata({ actionName: 'deleteReorderFromProductPage' })
+  .schema(z.object({productID: z.coerce.number() }))
+  .action(async ({ parsedInput: { productID }, ctx: { user, lang} }) => {
+    const { t } = await serverTranslation(lang, 'action-errors')
+	const currentLocation = await locationService.getLastVisited(user.id)
+	if (!currentLocation) {
+      throw new ActionError(t('restock-action.company-location-not-found'))
+	}
+
+    const existsLocation = await locationService.getByID(currentLocation)
+    if (!existsLocation) {
+      throw new ActionError(t('restock-action.company-location-not-found'))
+    }
+    if (existsLocation.customerID != user.customerID) {
+      throw new ActionError(
+        t('restock-action.company-location-belongs-to-your-company'),
+      )
+    }
+
+    const didDelete = await inventoryService.deleteReorderByIDs(
+      productID,
+	  currentLocation,
+      user.customerID,
+    )
+    if (!didDelete) {
+      throw new ActionError(t('minimum-stock-action.minimum-stock-not-deleted'))
+    }
+
+    revalidatePath(`/${lang}/varer/produkter/${productID}`)
+  })
